requirements.txt
fastapi
uvicorn
httpx
pydantic
langchain
langchain-groq
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
apscheduler
duckduckgo-search
python-dotenv
python-multipart
